moved tests to search-navigation
